+++ pyxel/inputs_outputs/parametric_outputs.py
@@ -11,7 +11,6 @@ import logging
 import operator
 import typing as t
 from pathlib import Path
-from time import strftime
 
 import h5py as h5
 import numpy as np
@@ -22,7 +21,7 @@ from astropy.io import fits as fits
 from pyxel import __version__ as version
 from pyxel.parametric import ParametricMode
 
-from .outputs import apply_run_number
+from .outputs import apply_run_number, create_output_directory
 
 if t.TYPE_CHECKING:
     from ..detectors import Detector
@@ -88,9 +87,7 @@ class ParametricOutputs:
 
         # self.user_plt_args = None  # type: t.Optional[PlotArguments]
         # self.save_parameter_to_file = save_parameter_to_file  # type: t.Optional[dict]
-        self.output_dir = (
-            Path(output_folder).joinpath("run_" + strftime("%Y%m%d_%H%M%S")).resolve()
-        )  # type: Path
+        self.output_dir = create_output_directory(output_folder)  # type: Path
 
         # TODO: Not related to a plot. Use by 'single' and 'parametric' modes.
         self.save_data_to_file = (
@@ -100,12 +97,6 @@ class ParametricOutputs:
         self.save_parametric_data = (
             save_parametric_data
         )  # type: t.Optional[t.Sequence[t.Mapping[str, t.Sequence[str]]]]
-
-        # TODO: reenable
-        # if self.output_dir.exists():
-        #    raise IsADirectoryError("Directory exists.")
-
-        self.output_dir.mkdir(parents=True, exist_ok=True)
 
         # self.default_ax_args = {
         #     "xlabel": None,